+++ app/actors/TranslateJavaSocketActor.java
@@ -1,4 +1,4 @@
-package models.actors;
+package actors;
 
 import akka.actor.ActorRef;
 import akka.actor.PoisonPill;
@@ -8,20 +8,20 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import play.libs.Json;
 
-public class VCSocketActor extends AbstractSocketActor {
+public class TranslateJavaSocketActor extends AbstractSocketActor {
 
-    public VCSocketActor(ActorRef out, String job, String project) {
+    public TranslateJavaSocketActor(ActorRef out, String job, String project) {
         super(out, job, project);
     }
 
     public static Props props(ActorRef out, String job, String project) {
         // http://doc.akka.io/docs/akka/snapshot/java/untyped-actors.html
-        return Props.create(new Creator<VCSocketActor>() {
+        return Props.create(new Creator<TranslateJavaSocketActor>() {
             private static final long serialVersionUID = 1L;
 
             @Override
-            public VCSocketActor create() throws Exception {
-                return new VCSocketActor(out, job, project);
+            public TranslateJavaSocketActor create() throws Exception {
+                return new TranslateJavaSocketActor(out, job, project);
             }
         });
     }
