+++ app/actors/VCSocketActor.java
@@ -1,27 +1,27 @@
-package models.actors;
+package actors;
 
 import akka.actor.ActorRef;
+import akka.actor.PoisonPill;
 import akka.actor.Props;
 import akka.japi.Creator;
 import com.fasterxml.jackson.databind.JsonNode;
-import controllers.compiler.TheoryAnalyzeInvoker;
-import java.util.HashMap;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 import play.libs.Json;
 
-public class AnalyzeSocketActor extends AbstractSocketActor {
+public class VCSocketActor extends AbstractSocketActor {
 
-    public AnalyzeSocketActor(ActorRef out, String job, String project) {
+    public VCSocketActor(ActorRef out, String job, String project) {
         super(out, job, project);
     }
 
     public static Props props(ActorRef out, String job, String project) {
         // http://doc.akka.io/docs/akka/snapshot/java/untyped-actors.html
-        return Props.create(new Creator<AnalyzeSocketActor>() {
+        return Props.create(new Creator<VCSocketActor>() {
             private static final long serialVersionUID = 1L;
 
             @Override
-            public AnalyzeSocketActor create() throws Exception {
-                return new AnalyzeSocketActor(out, job, project);
+            public VCSocketActor create() throws Exception {
+                return new VCSocketActor(out, job, project);
             }
         });
     }
@@ -32,9 +32,18 @@ public class AnalyzeSocketActor extends AbstractSocketActor {
             // Only deal with Strings
             if (message instanceof String) {
                 JsonNode request = Json.parse((String) message);
-                String[] args = {"-main", myWorkspacePath, "-webinterface", "Test.mt"};
-                TheoryAnalyzeInvoker invoker = new TheoryAnalyzeInvoker(args, myWebSocketOut);
-                invoker.executeJob(new HashMap<>());
+
+                // Create a JSON Object informing we are starting the job
+                ObjectNode result = Json.newObject();
+                result.put("status", "info");
+                result.put("msg", "Received request with the following parameters: "
+                        + myJob + " and " + myProject + ". Launching the RESOLVE compiler with the specified arguments.");
+
+                // Send the message through the websocket
+                myWebSocketOut.tell(result.toString(), self());
+
+                // Close the connection
+                self().tell(PoisonPill.getInstance(), self());
             }
             else {
                 // Send an error message back to user and close
