+++ app/controllers/CompilerSocket.java
@@ -1,9 +1,8 @@
 package controllers;
 
 import akka.stream.javadsl.Flow;
-import models.actors.*;
+import actors.*;
 import play.mvc.Controller;
-import play.mvc.Http;
 import play.mvc.LegacyWebSocket;
 import play.mvc.WebSocket;
